Moved a few files into correct directories
