--- app/src/main/java/ru/sedooj/cinemaandroidapp/network/cinema/CinemaNetworkRepository.kt
@@ -1,10 +1,0 @@
-package ru.sedooj.cinemaandroidapp.network.cinema
-
-import ru.sedooj.cinemaandroidapp.network.Client
-
-interface CinemaNetworkRepository : Client {
-
-    suspend fun getAllTodayFilms() : AllTodayFilmsOutput?
-
-    suspend fun getFilmById(input: GetFilmByIdInput): GetFilmByIdOutput?
-}
