--- app/src/main/java/ru/sedooj/cinemaandroidapp/network/cinema/GetFilmByIdInput.kt
@@ -1,5 +1,0 @@
-package ru.sedooj.cinemaandroidapp.network.cinema
-
-data class GetFilmByIdInput(
-    val id: Long
-)
