+++ app/src/main/java/ru/sedooj/cinemaandroidapp/network/cinema/film/GetFilmByIdInput.kt
@@ -1,0 +1,5 @@
+package ru.sedooj.cinemaandroidapp.network.cinema.film
+
+data class GetFilmByIdInput(
+    val id: Long
+)
